requirements.txt
paho-mqtt==1.3.0
requests==2.18.1
